ST-16-refactor: 화면 크기에 따른 반응형 네비게이션 개선
