+++ src/app/(user)/account/page.tsx
@@ -1,9 +1,5 @@
 function Account() {
-  return (
-    <div className="flex h-[1000px] w-full max-w-[800px] flex-col items-center justify-center bg-gray-400">
-      마이 페이지
-    </div>
-  );
+  return <div>마이 페이지</div>;
 }
 
 export default Account;
