+++ src/app/(user)/myactivities/page.tsx
@@ -1,9 +1,5 @@
 function Activities() {
-  return (
-    <div className="flex h-[1000px] w-full max-w-[800px] flex-col items-center justify-center bg-gray-400">
-      내 체험 관리
-    </div>
-  );
+  return <div>내 체험 관리</div>;
 }
 
 export default Activities;
