+++ src/app/(user)/reservation-status/page.tsx
@@ -1,9 +1,5 @@
 function ReservationStatus() {
-  return (
-    <div className="flex h-[1000px] w-full max-w-[800px] flex-col items-center justify-center bg-gray-400">
-      예약 현황
-    </div>
-  );
+  return <div>예약 현황</div>;
 }
 
 export default ReservationStatus;
