+++ src/app/(user)/reservations/page.tsx
@@ -1,9 +1,5 @@
 function Reservations() {
-  return (
-    <div className="flex h-[1000px] w-full max-w-[800px] flex-col items-center justify-center bg-gray-400">
-      예약 내역
-    </div>
-  );
+  return <div>예약 내역</div>;
 }
 
 export default Reservations;
